requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
python-decouple==3.8
celery==5.3.4
redis==5.0.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
